Testing automatic Docker build in CI [skip ci]

+++ tests/conftest.py
@@ -14,4 +14,5 @@ def test_dir():
         download_resources(test_dir=test_dir)
     yield test_dir
     logging.info(f"Removing the temporary directory for tests.")
-    shutil.rmtree(test_dir)
+    if os.path.exists(test_dir):
+        shutil.rmtree(test_dir)
